dev-cmd: rename `brew man` to `brew generate-man-completions`
