+++ Library/Homebrew/completions/fish.erb
@@ -9,7 +9,7 @@
 #     compare with already documented commands.
 # - For other changes: Edit this file.
 #
-# When done, regenerate the completions by running `brew man`.
+# When done, regenerate the completions by running `brew generate-man-completions`.
 %>
 # Fish shell completions for Homebrew
