+++ Library/Homebrew/manpages/brew.1.md.erb
@@ -11,7 +11,8 @@
 # - For environment variables: Edit `Library/Homebrew/env_config.rb`.
 # - For other changes: Edit this file.
 #
-# When done, regenerate the man page and its HTML version by running `brew man`.
+# When done, regenerate the man page and its HTML version by running
+# `brew generate-man-completions`.
 %>
 brew(1) -- The Missing Package Manager for macOS (or Linux)
 ===========================================================
