+++ Library/Homebrew/test/dev-cmd/generate-man-completions_spec.rb
@@ -1,0 +1,8 @@
+# typed: false
+# frozen_string_literal: true
+
+require "cmd/shared_examples/args_parse"
+
+describe "brew generate-man-completions" do
+  it_behaves_like "parseable arguments"
+end
